+++ eSTOL_Training_Tool/eSTOL_Training_Tool_Core/UI/FormCheckUpdate.Designer.cs
@@ -104,6 +104,7 @@
             Name = "FormCheckUpdate";
             ShowIcon = false;
             Text = "Check Update";
+            TopMost = true;
             ResumeLayout(false);
             PerformLayout();
         }